requirements.txt
Django == 1.11.3
django-debug-toolbar == 1.8
